# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation de l'API du module de billeterie automatique",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/all_users": {
            "get": {
                "tags": [
                    "The_user"
                ],
                "summary": "Afficher la liste des User",
                "description": "Renvoyer la liste de tous les utilisateurs ayant créer un compte ",
                "operationId": "304c8d1e31c6f13bf89c43b2473a629a",
                "responses": {
                    "200": {
                        "description": "Les utilisateurs ayant un compte ont été trouvé et leurs informations sont retournées",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Aucun utilisateur ayant créer un compte dans la base de données ",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user_show/{id}": {
            "get": {
                "tags": [
                    "The_user"
                ],
                "summary": "Afficher les infos d'un utilisateur",
                "description": "Renvoyer les infos du compte d'un utilisateur",
                "operationId": "10998f72ea862724428e95a35ef4a501",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Identifiant de l'étudiant à voir ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": " trouvé  !",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Aucun  ne correspond à l'identifiant saisi",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user_store": {
            "post": {
                "tags": [
                    "The_user"
                ],
                "summary": "Enregistrer un nouvel User",
                "operationId": "19b8caf736c08935d1eb21359da5cf07",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "poste",
                        "in": "query",
                        "description": "Poste occupé par l'utilisateur.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email de l'utilisateur.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": " enregistre avec succes.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Echec d'enregistrement de l'.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "402": {
                        "description": "Un utilisateur avec ces informations existe déja",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/user_destroy/{id}": {
            "delete": {
                "tags": [
                    "The_user"
                ],
                "summary": "Supprimer un compte d'utilisateur",
                "description": "Elle se charger de supprimer un  et donc son compte",
                "operationId": "263352c29ef8bee75f666ff3a6d68b84",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Identifiant de l'utilisateur à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": " supprimé avec succès!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "L'étudiant n'existe pas !",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    }
}